# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Management API",
        "version": "2"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/user/spreadsheet": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Geração dos dados dos usuários registrados em formato CSV",
                "operationId": "fc1920753236d546c52bf135fa5b8fe1",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "csv": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "csv": "name,cpf,email\\nRonaldo de Assis Moreira,16742019077,drnaoseioque@email.com\\n"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Importação de usuário através de arquivo CSV novo",
                "operationId": "6f6a2080a06eae2c2bf6a0289370d77c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "Campo do tipo arquivo, com o nome 'file', que recebe o arquivo CSV com os dados dos usuários",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "created_users": {
                                            "type": "string"
                                        },
                                        "date_time": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "created_users": 2,
                                        "date_time": "2023-12-28 04:10:10"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Erro interno: Spreadsheet error: line 2 | CPF already created"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/user/{id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Retorna os dados de um usuário específico por ID",
                "description": "Retorna os dados de um usuário específico por ID.",
                "operationId": "8e0284cc490469ddd7949f9446e2eec6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "cpf": {
                                            "type": "string"
                                        },
                                        "data_admissao": {
                                            "type": "string"
                                        },
                                        "company": {
                                            "type": "string"
                                        },
                                        "active": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Atualiza as informações (nome, CPF e email) de um usuário",
                "description": "Atualiza as informações (nome, CPF e email) de um usuário específico.",
                "operationId": "dd01c452fd878d777b377befec55bfb1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cpf",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Novo Nome"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678901"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "novo.email@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "cpf": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "User"
                ],
                "summary": "Realiza o soft delete de um usuário",
                "description": "Realiza o soft delete de um usuário.",
                "operationId": "2a1c56f3b5f18929b5bb0cbb19e2cd02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário a ser removido",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User soft deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User soft deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}